put Operand and its implementations in a separate package and renamed ExpressionQuestion to Question
