+++ back-end/question-generator/src/main/java/com/algebra/question_generator/controller/ExpressionController.java
@@ -1,11 +1,12 @@
 package com.algebra.question_generator.controller;
 
-import com.algebra.question_generator.model.ExpressionQuestion;
+import com.algebra.question_generator.model.Question;
 import com.algebra.question_generator.model.QuestionRequestDTO;
 import com.algebra.question_generator.service.ExpressionService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
@@ -17,9 +18,9 @@ public class ExpressionController {
   private ExpressionService expressionService;
 
   @GetMapping
-  public ResponseEntity<ExpressionQuestion> generateExpression(QuestionRequestDTO questionRequestDTO) {
+  public ResponseEntity<Question> generateExpression(@RequestBody QuestionRequestDTO questionRequestDTO) {
 
-    ExpressionQuestion response = expressionService.generateRandomExpression(questionRequestDTO);
+    Question response = expressionService.generateRandomExpression(questionRequestDTO);
     return ResponseEntity.ok().body(response);
   }
 }
