--- back-end/question-generator/src/main/java/com/algebra/question_generator/model/Operand.java
@@ -1,7 +1,0 @@
-package com.algebra.question_generator.model;
-
-public interface Operand {
-    Operand sum(Operand other);
-    Operand subtract(Operand other);
-    Operand multiply(Operand other);
-}
