+++ back-end/question-generator/src/main/java/com/algebra/question_generator/model/Operands/Operand.java
@@ -1,0 +1,9 @@
+package com.algebra.question_generator.model.Operands;
+
+public interface Operand {
+  Operand sum(Operand other);
+
+  Operand subtract(Operand other);
+
+  Operand multiply(Operand other);
+}
