+++ back-end/question-generator/src/main/java/com/algebra/question_generator/model/Operands/Rational.java
@@ -1,4 +1,4 @@
-package com.algebra.question_generator.model;
+package com.algebra.question_generator.model.Operands;
 
 import lombok.Getter;
 import lombok.Setter;
